Add / DataManager 및 Rest 로직 수정
